Rewrote POST controllers. Take the functionality outside the controllers.

+++ demo/src/main/java/com/example/demo/db/dao/OrderDAO.java
@@ -264,7 +264,7 @@ public class OrderDAO {
         try{
             logger.info("Start adding an items to orders");
             logger.info(order.getItems());
-            for (Map.Entry<Integer, Integer> entry:
+            for (Map.Entry<Long, Integer> entry:
                  order.getItems().entrySet()) {
                     logger.info(entry);
                     connection = DBManager.getInstance().getConnection();
